Update matrix output in GitHub workflow
This commit updates the way matrix output is handled in .github/workflows/action_testing.yml. Instead of using "::set-output", it now writes to the GITHUB_ENV. This is to ensure the environment variables are easily accessed throughout the GitHub actions workflow and prevent breaking of the pipeline in the future.