- fix: testing matrix generation in action_testing workflow
This commit updates the generation of the testing matrix in the GitHub Actions workflow 'action_testing'. The condition for populating the matrix has been changed to look for 'test-on:' prefix before the OS name. Additionally, it adds a check for empty matrix, and directly writes an empty array to the GITHUB_ENV in that case.

test-on: ubuntu-latest, windows-latest, macos-latest