small fixes (#165)
* - add: move Delta plotting functionality to class definition
- add: labeling for Delta plotting
- fix: include 'all' option when installing napari

* - add: table showing compatibility of astroCAST with different operating systems

* - add: improvements for GUI

* Refactor import statements and fix figure layout creation

In `cli_interfaces.py`, the helper module is imported directly where it's used instead of on top of the python file, improving code clarity and memory usage. The click argument's --save-path is also turned into a positional argument. In `preparation.py`, a condition is added to the figure layout creation to handle situations where there is only one pixel, ensuring an iterable object is always returned to support the subsequent code flow.

---------

Co-authored-by: Jan Philipp Reising <jan.reising@kise>