Update denoising test parameters and functions
Mainly, this commit focuses on altering the parameters and some function names within the test file pertaining to denoising. The most significant modification is in the `SubFrameDataset` class within the `denoising.py` file, where parameter names and defaults have been updated for clarification and alignment with standard conventions. This is to enhance code clarity and maintainability. The 'TestGenerators' class in the test file has also been renamed to 'TestDenoisingTensorflow'. Moreover, removed parts of the code which were no longer necessary (e.g., commented out or not reference by the system).

test-on:ubuntu-latest
test-on:windows-latest
test-on:macos-latest