Add 'max_per_file' key to 'inf_param' in denoising tests
The 'inf_param' dictionary within the test_denoising.py file now includes a 'max_per_file' key. This was added in three separate instances in the testing parameters dictionary. This will specify the maximum number of files that can be tested at a time, allowing for increased control over the testing process.