Update default values and error handling in denoising.py
In this update, the default values for 'allowed_rotation' and 'allowed_flip' parameters have been changed to a single int instead of a list. Additionally, error handling is improved by checking the type of 'self.mem_data' using isinstance instead of type, and by adding a ValueError when the path is not provided as a tif or h5 file. Some inconsistencies regarding 'chunk_size' value checking are also resolved.