Implemeted pytorch based denoising and updated tests (#166)
* - fix: small errors when testing Pytorch denoiser

- refactor: Refactor denoising code and remove unused UNet implementation:

The refactoring included simplifying the method call by removing redundant parameters. The unused implementation of UNet along with its related classes (InConv, Down, Up, OutConv) has been deleted. Additionally, some attributes were defined for further usage. Any batch normalization that is currently unimplemented raises a logging warning.

run-on: ubuntu-latest, windows-latest, macos-latest

* - increment version

test-on: ubuntu-latest windows-latest macos-latest

* - fix: testing matrix generation in action_testing workflow

This commit updates the generation of the testing matrix in the GitHub Actions workflow 'action_testing'. The condition for populating the matrix has been changed to look for 'test-on:' prefix before the OS name. Additionally, it adds a check for empty matrix, and directly writes an empty array to the GITHUB_ENV in that case.

* Update denoising module and tests with SubFrameDataset

The denoising module and its respective tests have been updated to utilize the new PyTorch class, SubFrameDataset. This has enabled a more streamlined and efficient data loading and pre-processing pipeline. Additionally, the changes include improved management of local and global normalization methods along with enhanced validation of acceptable input parameters. Also, the associated test suite has been updated to ensure all newly introduced functionalities are effectively tested.

* Update denoising test parameters and functions

Mainly, this commit focuses on altering the parameters and some function names within the test file pertaining to denoising. The most significant modification is in the `SubFrameDataset` class within the `denoising.py` file, where parameter names and defaults have been updated for clarification and alignment with standard conventions. This is to enhance code clarity and maintainability. The 'TestGenerators' class in the test file has also been renamed to 'TestDenoisingTensorflow'. Moreover, removed parts of the code which were no longer necessary (e.g., commented out or not reference by the system).

* Update action testing workflow to version 3

In this update, the name is changed to reflect the incremental version of the workflow. Additionally, the option for the log level in the pytest command has been corrected from "cli-log-level" to "log-cli-level".

* Update action testing workflow version

The action testing workflow version has been updated from v3 to v4. This change is reflected in the `action_testing.yml` configuration file.

* Update parameter types and fix parameter name

In denoising.SubFrameGenerator, the paths parameter now also accepts Path or List[Path]. Also in cli_interfaces.py, a parameter name was corrected from 'pre_post_frame' to 'pre_post_frames'. These changes improve flexibility and correct a naming error.

* Correct parameter in SubFrameGenerator instances and handle None type loc

This commit contains two primary changes. The first correction is made in the parameters of SubFrameGenerator instances, where 'pre_post_frame' is correctly renamed to 'pre_post_frames'. The second change is a modification to handle cases where 'loc' is None, assigning it an empty string value to avoid subsequent errors.

* Add 'max_per_file' key to 'inf_param' in denoising tests

The 'inf_param' dictionary within the test_denoising.py file now includes a 'max_per_file' key. This was added in three separate instances in the testing parameters dictionary. This will specify the maximum number of files that can be tested at a time, allowing for increased control over the testing process.

* Update installation documentation and dependency versions

The changes in this commit include an update to the installation documentation, specifically indicating the availability of Umap outlier preprocessing. Additionally, the version of the main package has been bumped up and the OS-specific versions for 'umap-learn' and 'tensorflow' dependencies have been generalized to non-specific versions.

* Add PyTorch support to model generation scripts

The `generate_pretrain_models.yaml` and `run_generate_models.py` scripts have been updated to include an option for using PyTorch in the model generation. This option can be toggled through config parameters `use_pytorch` and `use_cpu`. Also, the code has been restructured to conditionally generate training and validation data based on whether PyTorch is being used or not.

* Expand path input types in SubFrameDataset

The constructor for the class SubFrameDataset has been updated to accept either 'Path' or 'List[Path]' types in addition to string paths. This provides more flexibility in specifying file paths, accommodating for possibly more convenient or succinct path expressions.

* Update default values and error handling in denoising.py

In this update, the default values for 'allowed_rotation' and 'allowed_flip' parameters have been changed to a single int instead of a list. Additionally, error handling is improved by checking the type of 'self.mem_data' using isinstance instead of type, and by adding a ValueError when the path is not provided as a tif or h5 file. Some inconsistencies regarding 'chunk_size' value checking are also resolved.