Add chunk_strategy parameter to Video class
The chunk_strategy parameter has been added to the Video class' __init__ function and its IO.load calls. This optional parameter takes the values 'Z', 'balanced', or 'XY'; with a default of 'Z'. It allows for the configuration of how chunks of data are loaded and processed, enabling optimizations based on the specific use case.