Add peak detection method and improve video alignment
A new method 'find_peaks' has been added to detect signal peaks in a dataset and store the peak information in a data frame. In addition, the 'align' method has been updated to include extrapolation of timing data when aligning video frames with timing data. This addresses issues where the number of video frames exceed the length of the timing data.