Add new dev_sim.ipynb notebook
This commit includes a new notebook file, dev_sim.ipynb, which consists of code for modeling astrocast simulation. Although there's an AttributeError in this commit, the majority of the notebook contains code for setting up the environment and defining initial parameters for the simulation. Further updates would be necessary for bug fixes and improvements.