Update methods for handling molecule amounts and concentrations
This refactoring update improves the code's ability to calculate and manage the amounts and concentrations of molecules. Method names are updated for clarity. In addition, handling of ATP requirements and glutamate conversion processes are also optimized to more correctly represent physiological processes. This change will improve the accuracy and reliability of simulations.