Add placement and collision detection for astrocytes
The creation and placement of astrocytes in the simulation has been reworked. Instead of placing them at random points without any checks, the code now places astrocytes in a more controlled way, trying up to a maximum number of times if a random placement fails. Additionally, collision checks have also been put in place for the astrocytes and their branches.