Update astrocyte simulation model logic
The code changes reflect a substantial update to the logic of the astrocyte simulation model. These modifications include adjustments in the generation of random lines and hotspot distributions, introduction of astrocyte functions and attributes related to death and glutamate behavior, and numerous tweaks in attributes and parameters throughout the simulation and astrocyte classes.