Refactor molecule diffusion in simulation environment
The simulation environment has been refactored to make molecule diffusion across astrocytes and branches more accurate and efficient. Changes were made to the diffuse_molecules methods in the "Astrocyte" and "Branch" classes and a new navigate method was added to the "Compartment" class. These updates will lead to better performance of molecule diffusion in the Astrocast simulations.