Update script execution time in config and add dynamic time calculation
The changes made to the configuration yaml file reflect a decrease in set execution time for scripts and instead introduce a dynamic calculation of time in the cli_interfaces.py. The calculation is based on image size, with a different multiplier applied depending on the specific script. The changes aim to optimize scripting run-time.