Remove condition for ROI in time setting
The condition that excludes 'roi' from the time setting process in the cli_interfaces.py file has been removed. Previously, if a '/' was present in the "req_time", and the value was not 'roi', these values would be split. Now, this operation will be performed regardless of the 'roi' value.