Support Path objects for file handling
The changes in this commit support the use of Path objects in file handling operations across the cli_interfaces and preparation modules. Conversion from Path objects to strings is done when required for compatibility with functions expecting string paths. Furthermore, a change was made to utilize the visualize_h5_recursive function for a more thorough file visualization when a dataset cannot be found.