Update TIFF file loading for lazy mode
The previous TIFF file loading method was updated to improve efficiency. Now, if it's in lazy mode, the image will be read into a Dask delayed object stack first. This way, computations will be postponed until actually necessary, further enhancing the memory efficiency of the program. In the non-lazy path, the functionality remains the same.