Add size method and refactor video and event handling
Introduced a new method to calculate the size of data in the Video class. This takes into account whether it is a Dask or Numpy array. Also improved the handling of video and event creation in the analysis.py file, aiming at cleaner and more readable code. Also expanded the conditions to check if the 'video' is an instance of the 'Video' class.