Add figure description in manuscript
A detailed description for figure 3 in the manuscript has been added. The figure depicts a comparative visualization of a denoising algorithm's effects on video frames and pixel intensity measurements. It includes pre- and post- denoising frames and pixel intensity traces, as well as methodological details on the denoising algorithm such as training settings and test conditions.