Add Python style and command, update acronyms
A new Python style ('pyStyle') has been defined in the 'packages.tex' file and a new command for inline Python code ('\inlinepy') has been added in the 'custom_commands.tex' file. Additionally, in the 'frontiers.tex' file, an acronym for Recurring Neural Network (RNN) was added, and a typo in the Radial Basis Function (RBF) acronym was fixed.